Avoid recursive logger setup, and double log output
