+++ docling_serve/app.py
@@ -121,23 +121,6 @@ except (ImportError, Exception):
     pass
 
 
-# Set up custom logging as we'll be intermixes with FastAPI/Uvicorn's logging
-class ColoredLogFormatter(logging.Formatter):
-    COLOR_CODES = {
-        logging.DEBUG: "\033[94m",  # Blue
-        logging.INFO: "\033[92m",  # Green
-        logging.WARNING: "\033[93m",  # Yellow
-        logging.ERROR: "\033[91m",  # Red
-        logging.CRITICAL: "\033[95m",  # Magenta
-    }
-    RESET_CODE = "\033[0m"
-
-    def format(self, record):
-        color = self.COLOR_CODES.get(record.levelno, "")
-        record.levelname = f"{color}{record.levelname}{self.RESET_CODE}"
-        return super().format(record)
-
-
 # Configure logging based on settings
 # This will be called early, but can be reconfigured in __main__.py
 log_level = (
